// types.ts
export interface UserPrefs {
    travelMode: 'Private' | 'Metro' | 'Walk';
    dietary: string[];
}